Refactor to AktivitetskravVarselService
