+++ src/main/kotlin/no/nav/syfo/App.kt
@@ -6,9 +6,9 @@ import io.ktor.server.config.*
 import io.ktor.server.engine.*
 import io.ktor.server.netty.*
 import no.nav.syfo.aktivitetskrav.AktivitetskravService
+import no.nav.syfo.aktivitetskrav.AktivitetskravVarselService
 import no.nav.syfo.aktivitetskrav.database.AktivitetskravVarselRepository
-import no.nav.syfo.aktivitetskrav.kafka.AktivitetskravVurderingProducer
-import no.nav.syfo.aktivitetskrav.kafka.aktivitetskravVurderingProducerConfig
+import no.nav.syfo.aktivitetskrav.kafka.*
 import no.nav.syfo.application.ApplicationState
 import no.nav.syfo.application.Environment
 import no.nav.syfo.application.api.apiModule
@@ -16,6 +16,7 @@ import no.nav.syfo.application.cache.RedisStore
 import no.nav.syfo.application.cronjob.launchCronjobModule
 import no.nav.syfo.application.database.applicationDatabase
 import no.nav.syfo.application.database.databaseModule
+import no.nav.syfo.application.kafka.kafkaAivenProducerConfig
 import no.nav.syfo.application.kafka.launchKafkaModule
 import no.nav.syfo.client.azuread.AzureAdClient
 import no.nav.syfo.client.krr.KRRClient
@@ -80,8 +81,24 @@ fun main() {
             aktivitetskravVurderingProducerConfig(kafkaEnvironment = environment.kafka)
         )
     )
+    val arbeidstakervarselProducer = ArbeidstakervarselProducer(
+        kafkaArbeidstakervarselProducer = KafkaProducer(
+            kafkaAivenProducerConfig<KafkaArbeidstakervarselSerializer>(
+                kafkaEnvironment = environment.kafka,
+            )
+        )
+    )
+    val aktivitetskravVarselProducer = AktivitetskravVarselProducer(
+        kafkaProducer = KafkaProducer(
+            kafkaAivenProducerConfig<KafkaAktivitetskravVarselSerializer>(
+                kafkaEnvironment = environment.kafka,
+            )
+        )
+    )
+
     lateinit var aktivitetskravService: AktivitetskravService
     lateinit var aktivitetskravVarselRepository: AktivitetskravVarselRepository
+    lateinit var aktivitetskravVarselService: AktivitetskravVarselService
 
     val applicationEngineEnvironment = applicationEngineEnvironment {
         log = logger
@@ -96,9 +113,14 @@ fun main() {
             aktivitetskravVarselRepository = AktivitetskravVarselRepository(database = applicationDatabase)
             aktivitetskravService = AktivitetskravService(
                 aktivitetskravVurderingProducer = aktivitetskravVurderingProducer,
-                aktivitetskravVarselRepository = aktivitetskravVarselRepository,
                 database = applicationDatabase,
                 arenaCutoff = environment.arenaCutoff,
+            )
+            aktivitetskravVarselService = AktivitetskravVarselService(
+                aktivitetskravVarselRepository = aktivitetskravVarselRepository,
+                aktivitetskravVurderingProducer = aktivitetskravVurderingProducer,
+                arbeidstakervarselProducer = arbeidstakervarselProducer,
+                aktivitetskravVarselProducer = aktivitetskravVarselProducer,
                 pdfGenClient = pdfGenClient,
                 pdlClient = pdlClient,
                 krrClient = krrClient,
@@ -109,6 +131,7 @@ fun main() {
                 environment = environment,
                 wellKnownInternalAzureAD = wellKnownInternalAzureAD,
                 aktivitetskravService = aktivitetskravService,
+                aktivitetskravVarselService = aktivitetskravVarselService,
                 veilederTilgangskontrollClient = veilederTilgangskontrollClient,
             )
         }
@@ -128,8 +151,8 @@ fun main() {
             applicationState = applicationState,
             environment = environment,
             database = applicationDatabase,
-            aktivitetskravVarselRepository = aktivitetskravVarselRepository,
             aktivitetskravService = aktivitetskravService,
+            aktivitetskravVarselService = aktivitetskravVarselService,
             pdlClient = pdlClient,
             azureAdClient = azureAdClient,
         )
