+++ src/main/kotlin/no/nav/syfo/aktivitetskrav/AktivitetskravService.kt
@@ -1,14 +1,9 @@
 package no.nav.syfo.aktivitetskrav
 
-import no.nav.syfo.aktivitetskrav.api.ForhandsvarselDTO
 import no.nav.syfo.aktivitetskrav.database.*
 import no.nav.syfo.aktivitetskrav.domain.*
 import no.nav.syfo.aktivitetskrav.kafka.AktivitetskravVurderingProducer
 import no.nav.syfo.application.database.DatabaseInterface
-import no.nav.syfo.client.krr.KRRClient
-import no.nav.syfo.client.pdfgen.ForhandsvarselPdfDTO
-import no.nav.syfo.client.pdfgen.PdfGenClient
-import no.nav.syfo.client.pdl.PdlClient
 import no.nav.syfo.domain.PersonIdent
 import no.nav.syfo.oppfolgingstilfelle.domain.Oppfolgingstilfelle
 import java.sql.Connection
@@ -17,12 +12,8 @@ import java.util.*
 
 class AktivitetskravService(
     private val aktivitetskravVurderingProducer: AktivitetskravVurderingProducer,
-    private val aktivitetskravVarselRepository: AktivitetskravVarselRepository,
     private val database: DatabaseInterface,
     private val arenaCutoff: LocalDate,
-    private val pdfGenClient: PdfGenClient,
-    private val pdlClient: PdlClient,
-    private val krrClient: KRRClient,
 ) {
 
     internal fun createAktivitetskrav(
@@ -154,39 +145,5 @@ class AktivitetskravService(
         aktivitetskravVurderingProducer.sendAktivitetskravVurdering(
             aktivitetskrav = updatedAktivitetskrav
         )
-    }
-
-    suspend fun sendForhandsvarsel(
-        aktivitetskrav: Aktivitetskrav,
-        veilederIdent: String,
-        personIdent: PersonIdent,
-        forhandsvarselDTO: ForhandsvarselDTO,
-        callId: String,
-    ): AktivitetskravVarsel {
-        val personNavn = pdlClient.navn(personIdent)
-        val pdf = pdfGenClient.createForhandsvarselPdf(
-            callId,
-            ForhandsvarselPdfDTO.create(
-                mottakerNavn = personNavn,
-                mottakerFodselsnummer = personIdent.value,
-                documentComponents = forhandsvarselDTO.document,
-            )
-        )
-
-        val vurdering: AktivitetskravVurdering = forhandsvarselDTO.toAktivitetskravVurdering(veilederIdent)
-        val updatedAktivitetskrav = aktivitetskrav.vurder(aktivitetskravVurdering = vurdering)
-        val forhandsvarsel = AktivitetskravVarsel.create(forhandsvarselDTO.document)
-
-        val nyttForhandsvarsel = aktivitetskravVarselRepository.create(
-            aktivitetskrav = updatedAktivitetskrav,
-            varsel = forhandsvarsel,
-            pdf = pdf,
-        )
-
-        aktivitetskravVurderingProducer.sendAktivitetskravVurdering(
-            aktivitetskrav = updatedAktivitetskrav
-        )
-
-        return nyttForhandsvarsel.toAktivitetkravVarsel()
     }
 }
