+++ src/main/kotlin/no/nav/syfo/application/cronjob/CronjobModule.kt
@@ -1,27 +1,21 @@
 package no.nav.syfo.application.cronjob
 
 import no.nav.syfo.aktivitetskrav.AktivitetskravService
+import no.nav.syfo.aktivitetskrav.AktivitetskravVarselService
 import no.nav.syfo.aktivitetskrav.cronjob.*
-import no.nav.syfo.aktivitetskrav.kafka.KafkaArbeidstakervarselSerializer
-import no.nav.syfo.aktivitetskrav.database.AktivitetskravVarselRepository
-import no.nav.syfo.aktivitetskrav.kafka.AktivitetskravVarselProducer
 import no.nav.syfo.application.*
 import no.nav.syfo.application.database.DatabaseInterface
-import no.nav.syfo.application.kafka.kafkaAivenProducerConfig
-import no.nav.syfo.aktivitetskrav.kafka.ArbeidstakervarselProducer
-import no.nav.syfo.aktivitetskrav.kafka.KafkaAktivitetskravVarselSerializer
 import no.nav.syfo.client.azuread.AzureAdClient
 import no.nav.syfo.client.dokarkiv.DokarkivClient
 import no.nav.syfo.client.leaderelection.LeaderPodClient
 import no.nav.syfo.client.pdl.PdlClient
-import org.apache.kafka.clients.producer.KafkaProducer
 
 fun launchCronjobModule(
     applicationState: ApplicationState,
     environment: Environment,
     database: DatabaseInterface,
     aktivitetskravService: AktivitetskravService,
-    aktivitetskravVarselRepository: AktivitetskravVarselRepository,
+    aktivitetskravVarselService: AktivitetskravVarselService,
     pdlClient: PdlClient,
     azureAdClient: AzureAdClient,
 ) {
@@ -54,30 +48,14 @@ fun launchCronjobModule(
         dokarkivEnvironment = environment.clients.dokarkiv,
     )
     val journalforAktivitetskravVarselCronjob = JournalforAktivitetskravVarselCronjob(
-        aktivitetskravVarselRepository = aktivitetskravVarselRepository,
+        aktivitetskravVarselService = aktivitetskravVarselService,
         dokarkivClient = dokarkivClient,
         pdlClient = pdlClient
     )
     cronjobs.add(journalforAktivitetskravVarselCronjob)
 
-    val arbeidstakervarselProducer = ArbeidstakervarselProducer(
-        kafkaArbeidstakervarselProducer = KafkaProducer(
-            kafkaAivenProducerConfig<KafkaArbeidstakervarselSerializer>(
-                kafkaEnvironment = environment.kafka,
-            )
-        )
-    )
-    val aktivitetskravVarselProducer = AktivitetskravVarselProducer(
-        kafkaProducer = KafkaProducer(
-            kafkaAivenProducerConfig<KafkaAktivitetskravVarselSerializer>(
-                kafkaEnvironment = environment.kafka,
-            )
-        )
-    )
     val publiserAktivitetskravVarselCronjob = PubliserAktivitetskravVarselCronjob(
-        aktivitetskravVarselRepository = aktivitetskravVarselRepository,
-        arbeidstakervarselProducer = arbeidstakervarselProducer,
-        aktivitetskravVarselProducer = aktivitetskravVarselProducer,
+        aktivitetskravVarselService = aktivitetskravVarselService,
     )
     cronjobs.add(publiserAktivitetskravVarselCronjob)
