+++ src/test/kotlin/no/nav/syfo/testhelper/TestApiModule.kt
@@ -1,7 +1,9 @@
 package no.nav.syfo.testhelper
 
 import io.ktor.server.application.*
+import io.mockk.mockk
 import no.nav.syfo.aktivitetskrav.AktivitetskravService
+import no.nav.syfo.aktivitetskrav.AktivitetskravVarselService
 import no.nav.syfo.aktivitetskrav.database.AktivitetskravVarselRepository
 import no.nav.syfo.aktivitetskrav.kafka.AktivitetskravVurderingProducer
 import no.nav.syfo.application.api.apiModule
@@ -15,12 +17,17 @@ fun Application.testApiModule(
         aktivitetskravVurderingProducer = aktivitetskravVurderingProducer,
         database = externalMockEnvironment.database,
         arenaCutoff = externalMockEnvironment.environment.arenaCutoff,
+    )
+    val aktivitetskravVarselService = AktivitetskravVarselService(
         pdfGenClient = externalMockEnvironment.pdfgenClient,
         aktivitetskravVarselRepository = AktivitetskravVarselRepository(
             database = externalMockEnvironment.database
         ),
         pdlClient = externalMockEnvironment.pdlClient,
         krrClient = externalMockEnvironment.krrClient,
+        arbeidstakervarselProducer = mockk(),
+        aktivitetskravVarselProducer = mockk(),
+        aktivitetskravVurderingProducer = aktivitetskravVurderingProducer,
     )
     val veilederTilgangskontrollClient = VeilederTilgangskontrollClient(
         azureAdClient = externalMockEnvironment.azureAdClient,
@@ -33,6 +40,7 @@ fun Application.testApiModule(
         environment = externalMockEnvironment.environment,
         wellKnownInternalAzureAD = externalMockEnvironment.wellKnownInternalAzureAD,
         aktivitetskravService = aktivitetskravService,
+        aktivitetskravVarselService = aktivitetskravVarselService,
         veilederTilgangskontrollClient = veilederTilgangskontrollClient,
     )
 }
